- Modified npc print levels for npc client to be in   range 1-5 for events not happening to frequent,   6-10 for printing on advance and 11-15 for very detailed   printings.

+++ src/common/paws/pawswidget.cpp
@@ -1846,7 +1846,7 @@ void pawsWidget::Resize( int deltaX, int deltaY, int flags )
     {
         int delta = (deltaX<deltaY) ? deltaX:deltaY;
         newFrame.ymax += delta;
-        newFrame.xmax += delta*scale;
+        newFrame.xmax += (int)(delta*scale);
     }
 
     if ( flags & RESIZE_LEFT)
